+++ src/Apps/Artwork/Components/ArtworkLightbox.tsx
@@ -5,7 +5,7 @@ import * as React from "react"
 import { Link } from "react-head"
 import { createFragmentContainer, graphql } from "react-relay"
 import { useSystemContext } from "System"
-import { useLocalImage } from "Utils/localImagesHelpers"
+import { useLocalImage } from "Utils/localImageHelpers"
 import { userIsTeam } from "Utils/user"
 import { ArtworkLightbox_artwork$data } from "__generated__/ArtworkLightbox_artwork.graphql"
 import { ArtworkLightboxPlaceholder } from "./ArtworkLightboxPlaceholder"
