+++ src/Apps/MyCollection/Routes/EditArtwork/MyCollectionCreateArtwork.tsx
@@ -14,7 +14,7 @@ import { useEffect, useState } from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import { useRouter } from "System/Router/useRouter"
 import { useFeatureFlag } from "System/useFeatureFlag"
-import { storeLocalImage } from "Utils/localImagesHelpers"
+import { storeLocalImage } from "Utils/localImageHelpers"
 import createLogger from "Utils/logger"
 import { MyCollectionCreateArtwork_me$data } from "__generated__/MyCollectionCreateArtwork_me.graphql"
 import { getMyCollectionArtworkFormInitialValues } from "./Utils/artworkFormHelpers"
