+++ src/Apps/MyCollection/Routes/EditArtwork/MyCollectionEditArtwork.tsx
@@ -12,7 +12,7 @@ import { reverse } from "lodash"
 import { createFragmentContainer, graphql } from "react-relay"
 import { useRouter } from "System/Router/useRouter"
 import { useFeatureFlag } from "System/useFeatureFlag"
-import { storeLocalImage } from "Utils/localImagesHelpers"
+import { storeLocalImage } from "Utils/localImageHelpers"
 import createLogger from "Utils/logger"
 import { wait } from "Utils/wait"
 import { MyCollectionEditArtwork_artwork$data } from "__generated__/MyCollectionEditArtwork_artwork.graphql"
