+++ src/Utils/localImageHelpers.ts
